+++ Ecs/include/TRA/ecs/chunk.hpp
@@ -8,8 +8,8 @@ namespace tra::ecs
 	struct Chunk
 	{
 		uint8_t* m_data;
-		uint16_t m_count;
-		uint16_t m_capacity;
+		size_t m_count;
+		size_t m_capacity;
 
 		const ChunkLayout* m_layout;
 	};
