+++ Ecs/include/TRA/ecs/componentLibrary.hpp
@@ -15,9 +15,9 @@ namespace tra::ecs
 	{
 	public:
 		template<typename T>
-		static uint8_t registerComponent()
+		static size_t registerComponent()
 		{
-			uint8_t id = static_cast<uint8_t>(m_components.size());
+			size_t id = m_components.size();
 
 			ComponentInfo info;
 			info.m_size = sizeof(T);
@@ -51,7 +51,7 @@ namespace tra::ecs
 			return m_components.at(m_componentLookUp.at(typeid(T)));
 		}
 
-		TRA_API static const ComponentInfo& get(uint8_t _id);
+		TRA_API static const ComponentInfo& get(size_t _id);
 
 	private:
 		TRA_API static inline std::vector<ComponentInfo> m_components;
